requirements.txt
streamlit
pdfplumber
pillow
pytesseract
pandas
openpyxl
SQLAlchemy
psycopg2-binary